[#12] Adding Wix installer for Windows
