+++ FileSorter.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32228.430
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "FileSorter", "FileSorter\FileSorter.pyproj", "{A93E03A7-DBB8-4B5D-AE9A-D560CAE9D734}"
+EndProject
+Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "FileSorterInstaller", "FileSorterInstaller\FileSorterInstaller.wixproj", "{26CC829B-F68D-46A4-9E01-F4B3DA922787}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A93E03A7-DBB8-4B5D-AE9A-D560CAE9D734}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A93E03A7-DBB8-4B5D-AE9A-D560CAE9D734}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A93E03A7-DBB8-4B5D-AE9A-D560CAE9D734}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A93E03A7-DBB8-4B5D-AE9A-D560CAE9D734}.Release|x86.ActiveCfg = Release|Any CPU
+		{26CC829B-F68D-46A4-9E01-F4B3DA922787}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{26CC829B-F68D-46A4-9E01-F4B3DA922787}.Debug|Any CPU.Build.0 = Debug|x86
+		{26CC829B-F68D-46A4-9E01-F4B3DA922787}.Debug|x86.ActiveCfg = Debug|x86
+		{26CC829B-F68D-46A4-9E01-F4B3DA922787}.Debug|x86.Build.0 = Debug|x86
+		{26CC829B-F68D-46A4-9E01-F4B3DA922787}.Release|Any CPU.ActiveCfg = Release|x86
+		{26CC829B-F68D-46A4-9E01-F4B3DA922787}.Release|Any CPU.Build.0 = Release|x86
+		{26CC829B-F68D-46A4-9E01-F4B3DA922787}.Release|x86.ActiveCfg = Release|x86
+		{26CC829B-F68D-46A4-9E01-F4B3DA922787}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E2532FAA-3FC1-481D-8D3E-7F72A78F0890}
+	EndGlobalSection
+EndGlobal
