+++ FileSorter/FileSorter.pyproj
@@ -1,0 +1,47 @@
+<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="4.0">
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <SchemaVersion>2.0</SchemaVersion>
+    <ProjectGuid>a93e03a7-dbb8-4b5d-ae9a-d560cae9d734</ProjectGuid>
+    <ProjectHome>.</ProjectHome>
+    <StartupFile>tools\ReleaseTools\builldExecutables.py</StartupFile>
+    <SearchPath>
+    </SearchPath>
+    <WorkingDirectory>.</WorkingDirectory>
+    <OutputPath>.</OutputPath>
+    <Name>FileSorter</Name>
+    <RootNamespace>FileSorter</RootNamespace>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
+    <DebugSymbols>true</DebugSymbols>
+    <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
+    <DebugSymbols>true</DebugSymbols>
+    <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="tools\ReleaseTools\builldExecutables.py" />
+    <Compile Include="fileSorterCLI.py" />
+    <Compile Include="fileSorterGUI.py" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="tools\" />
+    <Folder Include="tools\ReleaseTools\" />
+    <Folder Include="tools\ReleaseTools\icons\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="tools\ReleaseTools\icons\cli.ico" />
+    <Content Include="tools\ReleaseTools\icons\gui.ico" />
+    <Content Include="tools\ReleaseTools\versionFile.txt" />
+  </ItemGroup>
+  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets" />
+  <!-- Uncomment the CoreCompile target to enable the Build command in
+       Visual Studio and specify your pre- and post-build commands in
+       the BeforeBuild and AfterBuild targets below. -->
+  <!--<Target Name="CoreCompile" />-->
+  <Target Name="BeforeBuild">
+  </Target>
+  <Target Name="AfterBuild">
+  </Target>
+</Project>
