+++ FileSorter/tools/ReleaseTools/versionFile.txt
@@ -1,0 +1,28 @@
+VSVersionInfo(
+  ffi=FixedFileInfo(
+    filevers=(1, 1, 0, 0),
+    prodvers=(1, 1, 0, 0),
+    mask=0x3f,
+    flags=0x0,
+    OS=0x40004,
+    fileType=0x1,
+    subtype=0x0,
+    date=(0, 0)
+    ),
+  kids=[
+    StringFileInfo(
+      [
+      StringTable(
+        u'040904B0',
+        [StringStruct(u'CompanyName', u'Jose Juan Jaramillo Polo'),
+        StringStruct(u'FileDescription', u'FileSorter'),
+        StringStruct(u'FileVersion', u'1.1.0.0'),
+        StringStruct(u'InternalName', u'FileSorter'),
+        StringStruct(u'LegalCopyright', u'\xa9 GNU'),
+        StringStruct(u'OriginalFilename', u'FileSorter.Exe'),
+        StringStruct(u'ProductName', u'FileSorter'),
+        StringStruct(u'ProductVersion', u'1.1.0.0')])
+      ]), 
+    VarFileInfo([VarStruct(u'Translation', [1033, 1200])])
+  ]
+)
